+++ NEWS.md
@@ -34,6 +34,8 @@
 
 - `cutData()` gains the `wd.res` argument, which can take one of `4`, `8`, or `16`, defaulting to `8`. `4` cuts the data into North, East, South and West. `16` cuts the data into N, NNE, NE, ENE, etc. All plotting function's `type` argument now responds to this, showing only four panels when `wd.res = 4` and a 5x5 grid of sixteen panels when `wd.res = 16`.
 
+-  `TheilSen` and `smoothTrend` now use a more straightfoward way to imput missing data when `deseason = TRUE` and missing monthly data are present based on monthly linear regression by month. The user is alerted to the imputation and the monthly plot shows the imputed data as a filled grey circle.
+
 ## Bug Fixes
 
 - `timeAverage()` now has a more robust approach to multi-period averaging time such as "3 day". Bin boundaries are at fixed periods of time, which should ensure consistency across data sets that start at different times. This also means there is less need to use the argument `start.date` unless there is a need to extend the start date of the data for some reason e.g. to the beginning of a year. THis change may result in slight differences in returned output but should not affect periods such as "day", "hour" and "month".